refactor(models): move SwComponentType and Operation InstanceRefs to Components package (Rule 0007 A)
