+++ src/armodel/models/M2/AUTOSARTemplates/SWComponentTemplate/Composition/InstanceRefs.py
@@ -126,121 +126,6 @@ class RPortInCompositionInstanceRef(PortInCompositionTypeInstanceRef):
         return self
 
 
-class OperationInAtomicSwcInstanceRef(AtpInstanceRef, ABC):
-    """
-    Abstract base class for operation instance references within an atomic
-    software component type.
-    """
-
-    # OperationInAtomicSwcInstanceRef method parity checklist:
-    # [ ] __init__                     [x] impl  [ ] docstring  [ ] test
-    # [ ] getBaseRef                   [x] impl  [ ] docstring  [ ] test
-    # [ ] setBaseRef                   [x] impl  [ ] docstring  [ ] test
-    # [ ] getContextPortRef            [x] impl  [ ] docstring  [ ] test
-    # [ ] setContextPortRef            [x] impl  [ ] docstring  [ ] test
-    # [ ] getTargetOperationRef        [x] impl  [ ] docstring  [ ] test
-    # [ ] setTargetOperationRef        [x] impl  [ ] docstring  [ ] test
-
-    def __init__(self):
-        if type(self) is OperationInAtomicSwcInstanceRef:
-            raise TypeError("OperationInAtomicSwcInstanceRef is an abstract class.")
-
-        super().__init__()
-
-        self.baseRef: RefType = None
-        self.contextPortRef: RefType = None
-        self.targetOperationRef: RefType = None
-
-    def getBaseRef(self):
-        return self.baseRef
-
-    def setBaseRef(self, value):
-        self.baseRef = value
-        return self
-
-    def getContextPortRef(self):
-        return self.contextPortRef
-
-    def setContextPortRef(self, value):
-        self.contextPortRef = value
-        return self
-
-    def getTargetOperationRef(self):
-        return self.targetOperationRef
-
-    def setTargetOperationRef(self, value):
-        self.targetOperationRef = value
-        return self
-
-
-class POperationInAtomicSwcInstanceRef(OperationInAtomicSwcInstanceRef):
-    """
-    Instance reference to a provided operation in an atomic software
-    component through a PPortPrototype.
-    """
-
-    # POperationInAtomicSwcInstanceRef method parity checklist:
-    # [ ] __init__                     [x] impl  [ ] docstring  [ ] test
-    # [ ] getContextPPortRef           [x] impl  [ ] docstring  [ ] test
-    # [ ] setContextPPortRef           [x] impl  [ ] docstring  [ ] test
-    # [ ] getTargetProvidedOperationRef [x] impl  [ ] docstring  [ ] test
-    # [ ] setTargetProvidedOperationRef [x] impl  [ ] docstring  [ ] test
-
-    def __init__(self):
-        super().__init__()
-
-        self.contextPPortRef: RefType = None
-        self.targetProvidedOperationRef: RefType = None
-
-    def getContextPPortRef(self):
-        return self.contextPPortRef
-
-    def setContextPPortRef(self, value):
-        self.contextPPortRef = value
-        return self
-
-    def getTargetProvidedOperationRef(self):
-        return self.targetProvidedOperationRef
-
-    def setTargetProvidedOperationRef(self, value):
-        self.targetProvidedOperationRef = value
-        return self
-
-
-class ROperationInAtomicSwcInstanceRef(OperationInAtomicSwcInstanceRef):
-    """
-    Instance reference to a required operation in an atomic software
-    component through an RPortPrototype.
-    """
-
-    # ROperationInAtomicSwcInstanceRef method parity checklist:
-    # [ ] __init__                     [x] impl  [ ] docstring  [ ] test
-    # [ ] getContextRPortRef           [x] impl  [ ] docstring  [ ] test
-    # [ ] setContextRPortRef           [x] impl  [ ] docstring  [ ] test
-    # [ ] getTargetRequiredOperationRef [x] impl  [ ] docstring  [ ] test
-    # [ ] setTargetRequiredOperationRef [x] impl  [ ] docstring  [ ] test
-
-    def __init__(self):
-        super().__init__()
-
-        self.contextRPortRef: RefType = None
-        self.targetRequiredOperationRef: RefType = None
-
-    def getContextRPortRef(self):
-        return self.contextRPortRef
-
-    def setContextRPortRef(self, value):
-        self.contextRPortRef = value
-        return self
-
-    def getTargetRequiredOperationRef(self):
-        return self.targetRequiredOperationRef
-
-    def setTargetRequiredOperationRef(self, value):
-        self.targetRequiredOperationRef = value
-        return self
-
-
 class InstanceEventInCompositionInstanceRef(AtpInstanceRef):
     """
     Instance reference to an RTEEvent in the context of a CompositionSwComponentType.
