+++ tests/test_armodel/models/M2/AUTOSARTemplates/SWComponentTemplate/Components/test_InstanceRefs.py
@@ -9,10 +9,13 @@ from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.
 from armodel.models.M2.AUTOSARTemplates.SWComponentTemplate.Components.InstanceRefs import (
     InnerPortGroupInCompositionInstanceRef,
     ModeGroupInAtomicSwcInstanceRef,
+    OperationInAtomicSwcInstanceRef,
     PModeGroupInAtomicSwcInstanceRef,
+    POperationInAtomicSwcInstanceRef,
     PTriggerInAtomicSwcTypeInstanceRef,
     RModeGroupInAtomicSWCInstanceRef,
     RModeInAtomicSwcInstanceRef,
+    ROperationInAtomicSwcInstanceRef,
     RVariableInAtomicSwcInstanceRef,
     TriggerInAtomicSwcInstanceRef,
     VariableInAtomicSwcInstanceRef,
@@ -307,3 +310,100 @@ class TestInnerPortGroupInCompositionInstanceRef:
         target_ref.setValue("/Target/Ref")
         instance_ref.setTargetRef(target_ref)
         assert instance_ref.getTargetRef() == target_ref
+
+
+class TestOperationInAtomicSwcInstanceRef:
+    """Test class for OperationInAtomicSwcInstanceRef abstract class."""
+
+    def test_operation_in_atomic_swc_instance_ref_abstract(self):
+        """Test that OperationInAtomicSwcInstanceRef is an abstract class."""
+        with pytest.raises(TypeError):
+            OperationInAtomicSwcInstanceRef()
+
+
+class TestPOperationInAtomicSwcInstanceRef:
+    """Test class for POperationInAtomicSwcInstanceRef class."""
+
+    def test_p_operation_in_atomic_swc_instance_ref_initialization(self):
+        """Test POperationInAtomicSwcInstanceRef initialization and methods."""
+        instance_ref = POperationInAtomicSwcInstanceRef()
+
+        assert instance_ref.baseRef is None
+        assert instance_ref.contextPortRef is None
+        assert instance_ref.targetOperationRef is None
+        assert instance_ref.contextPPortRef is None
+        assert instance_ref.targetProvidedOperationRef is None
+
+        # Test baseRef methods
+        base_ref = RefType()
+        base_ref.setValue("/Base/Ref")
+        instance_ref.setBaseRef(base_ref)
+        assert instance_ref.getBaseRef() == base_ref
+
+        # Test contextPortRef methods
+        context_port_ref = RefType()
+        context_port_ref.setValue("/Context/Port")
+        instance_ref.setContextPortRef(context_port_ref)
+        assert instance_ref.getContextPortRef() == context_port_ref
+
+        # Test targetOperationRef methods
+        target_operation_ref = RefType()
+        target_operation_ref.setValue("/Target/Operation")
+        instance_ref.setTargetOperationRef(target_operation_ref)
+        assert instance_ref.getTargetOperationRef() == target_operation_ref
+
+        # Test contextPPortRef methods
+        context_p_port_ref = RefType()
+        context_p_port_ref.setValue("/Context/P/Port")
+        instance_ref.setContextPPortRef(context_p_port_ref)
+        assert instance_ref.getContextPPortRef() == context_p_port_ref
+
+        # Test targetProvidedOperationRef methods
+        target_provided_operation_ref = RefType()
+        target_provided_operation_ref.setValue("/Target/Provided/Operation")
+        instance_ref.setTargetProvidedOperationRef(target_provided_operation_ref)
+        assert instance_ref.getTargetProvidedOperationRef() == target_provided_operation_ref
+
+
+class TestROperationInAtomicSwcInstanceRef:
+    """Test class for ROperationInAtomicSwcInstanceRef class."""
+
+    def test_r_operation_in_atomic_swc_instance_ref_initialization(self):
+        """Test ROperationInAtomicSwcInstanceRef initialization and methods."""
+        instance_ref = ROperationInAtomicSwcInstanceRef()
+
+        assert instance_ref.baseRef is None
+        assert instance_ref.contextPortRef is None
+        assert instance_ref.targetOperationRef is None
+        assert instance_ref.contextRPortRef is None
+        assert instance_ref.targetRequiredOperationRef is None
+
+        # Test baseRef methods
+        base_ref = RefType()
+        base_ref.setValue("/Base/Ref")
+        instance_ref.setBaseRef(base_ref)
+        assert instance_ref.getBaseRef() == base_ref
+
+        # Test contextPortRef methods
+        context_port_ref = RefType()
+        context_port_ref.setValue("/Context/Port")
+        instance_ref.setContextPortRef(context_port_ref)
+        assert instance_ref.getContextPortRef() == context_port_ref
+
+        # Test targetOperationRef methods
+        target_operation_ref = RefType()
+        target_operation_ref.setValue("/Target/Operation")
+        instance_ref.setTargetOperationRef(target_operation_ref)
+        assert instance_ref.getTargetOperationRef() == target_operation_ref
+
+        # Test contextRPortRef methods
+        context_r_port_ref = RefType()
+        context_r_port_ref.setValue("/Context/R/Port")
+        instance_ref.setContextRPortRef(context_r_port_ref)
+        assert instance_ref.getContextRPortRef() == context_r_port_ref
+
+        # Test targetRequiredOperationRef methods
+        target_required_operation_ref = RefType()
+        target_required_operation_ref.setValue("/Target/Required/Operation")
+        instance_ref.setTargetRequiredOperationRef(target_required_operation_ref)
+        assert instance_ref.getTargetRequiredOperationRef() == target_required_operation_ref
