+++ tests/test_armodel/models/M2/AUTOSARTemplates/SWComponentTemplate/Composition/test_InstanceRefs.py
@@ -8,11 +8,8 @@ import pytest
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.PrimitiveTypes import RefType
 from armodel.models.M2.AUTOSARTemplates.SWComponentTemplate.Composition.InstanceRefs import (
     InstanceEventInCompositionInstanceRef,
-    OperationInAtomicSwcInstanceRef,
-    POperationInAtomicSwcInstanceRef,
     PortInCompositionTypeInstanceRef,
     PPortInCompositionInstanceRef,
-    ROperationInAtomicSwcInstanceRef,
     RPortInCompositionInstanceRef,
 )
 
@@ -112,103 +109,6 @@ class TestRPortInCompositionInstanceRef:
         target_r_port_ref.setValue("/Target/R/Port")
         instance_ref.setTargetRPortRef(target_r_port_ref)
         assert instance_ref.getTargetRPortRef() == target_r_port_ref
-
-
-class TestOperationInAtomicSwcInstanceRef:
-    """Test class for OperationInAtomicSwcInstanceRef abstract class."""
-
-    def test_operation_in_atomic_swc_instance_ref_abstract(self):
-        """Test that OperationInAtomicSwcInstanceRef is an abstract class."""
-        with pytest.raises(TypeError):
-            OperationInAtomicSwcInstanceRef()
-
-
-class TestPOperationInAtomicSwcInstanceRef:
-    """Test class for POperationInAtomicSwcInstanceRef class."""
-
-    def test_p_operation_in_atomic_swc_instance_ref_initialization(self):
-        """Test POperationInAtomicSwcInstanceRef initialization and methods."""
-        instance_ref = POperationInAtomicSwcInstanceRef()
-
-        assert instance_ref.baseRef is None
-        assert instance_ref.contextPortRef is None
-        assert instance_ref.targetOperationRef is None
-        assert instance_ref.contextPPortRef is None
-        assert instance_ref.targetProvidedOperationRef is None
-
-        # Test baseRef methods
-        base_ref = RefType()
-        base_ref.setValue("/Base/Ref")
-        instance_ref.setBaseRef(base_ref)
-        assert instance_ref.getBaseRef() == base_ref
-
-        # Test contextPortRef methods
-        context_port_ref = RefType()
-        context_port_ref.setValue("/Context/Port")
-        instance_ref.setContextPortRef(context_port_ref)
-        assert instance_ref.getContextPortRef() == context_port_ref
-
-        # Test targetOperationRef methods
-        target_operation_ref = RefType()
-        target_operation_ref.setValue("/Target/Operation")
-        instance_ref.setTargetOperationRef(target_operation_ref)
-        assert instance_ref.getTargetOperationRef() == target_operation_ref
-
-        # Test contextPPortRef methods
-        context_p_port_ref = RefType()
-        context_p_port_ref.setValue("/Context/P/Port")
-        instance_ref.setContextPPortRef(context_p_port_ref)
-        assert instance_ref.getContextPPortRef() == context_p_port_ref
-
-        # Test targetProvidedOperationRef methods
-        target_provided_operation_ref = RefType()
-        target_provided_operation_ref.setValue("/Target/Provided/Operation")
-        instance_ref.setTargetProvidedOperationRef(target_provided_operation_ref)
-        assert instance_ref.getTargetProvidedOperationRef() == target_provided_operation_ref
-
-
-class TestROperationInAtomicSwcInstanceRef:
-    """Test class for ROperationInAtomicSwcInstanceRef class."""
-
-    def test_r_operation_in_atomic_swc_instance_ref_initialization(self):
-        """Test ROperationInAtomicSwcInstanceRef initialization and methods."""
-        instance_ref = ROperationInAtomicSwcInstanceRef()
-
-        assert instance_ref.baseRef is None
-        assert instance_ref.contextPortRef is None
-        assert instance_ref.targetOperationRef is None
-        assert instance_ref.contextRPortRef is None
-        assert instance_ref.targetRequiredOperationRef is None
-
-        # Test baseRef methods
-        base_ref = RefType()
-        base_ref.setValue("/Base/Ref")
-        instance_ref.setBaseRef(base_ref)
-        assert instance_ref.getBaseRef() == base_ref
-
-        # Test contextPortRef methods
-        context_port_ref = RefType()
-        context_port_ref.setValue("/Context/Port")
-        instance_ref.setContextPortRef(context_port_ref)
-        assert instance_ref.getContextPortRef() == context_port_ref
-
-        # Test targetOperationRef methods
-        target_operation_ref = RefType()
-        target_operation_ref.setValue("/Target/Operation")
-        instance_ref.setTargetOperationRef(target_operation_ref)
-        assert instance_ref.getTargetOperationRef() == target_operation_ref
-
-        # Test contextRPortRef methods
-        context_r_port_ref = RefType()
-        context_r_port_ref.setValue("/Context/R/Port")
-        instance_ref.setContextRPortRef(context_r_port_ref)
-        assert instance_ref.getContextRPortRef() == context_r_port_ref
-
-        # Test targetRequiredOperationRef methods
-        target_required_operation_ref = RefType()
-        target_required_operation_ref.setValue("/Target/Required/Operation")
-        instance_ref.setTargetRequiredOperationRef(target_required_operation_ref)
-        assert instance_ref.getTargetRequiredOperationRef() == target_required_operation_ref
 
 
 class TestInstanceEventInCompositionInstanceRef:
